zaml t

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: otait-ta <otait-ta@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 12:31:59 by otait-ta          #+#    #+#             */
-/*   Updated: 2022/10/17 11:44:02 by otait-ta         ###   ########.fr       */
+/*   Updated: 2022/10/17 18:07:29 by otait-ta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *b, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	if (!b)
 		return ;
