+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: otait-ta <otait-ta@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/10 11:15:09 by otait-ta          #+#    #+#             */
-/*   Updated: 2022/10/17 11:47:38 by otait-ta         ###   ########.fr       */
+/*   Updated: 2022/10/17 18:06:41 by otait-ta         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned char	*__s1;
-	unsigned char	*__s2;
-	int				i;
+	unsigned char		*__s1;
+	unsigned char		*__s2;
+	size_t				i;
 
 	i = 0;
 	__s1 = ((unsigned char *)s1);
